fix(routines): routine screen's trigger chip times out instead of spinning forever (PRODUCT-1772)
RoutineActivationChip read the raw per-agent trigger status and mapped an
absent item to "Checking your trigger…" with no timeout, unlike the grid's row
badges. A moved agent's webhook routine (whose trigger rows the backend could
not see, fixed in cloud) therefore spun across restarts with no way to mint
the address.

The verification timeout is extracted from the grid's view model into
useTriggerStatusTimeouts and the chip resolves its status through it, so any
routine the host never reports on ends in the existing "couldn't confirm"
error after 45s. One copy of the rule, both surfaces bound to it.
