+++ app/src/components/agent/trigger-status-view-model.ts
@@ -1,6 +1,6 @@
 import type { Routine } from "@houston-ai/engine-client";
 import type { TriggerStatusItem } from "@houston-ai/routines";
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useCallback, useMemo } from "react";
 import { useTranslation } from "react-i18next";
 import { useIntegrationToolkits } from "../../hooks/queries/use-integrations";
 import { useCapabilities } from "../../hooks/use-capabilities";
@@ -8,13 +8,10 @@ import { useUIStore } from "../../stores/ui";
 import { INTEGRATION_PROVIDER } from "../integrations/model";
 import { INTEGRATIONS_VIEW_ID } from "../integrations-view/id";
 import {
-  TRIGGER_STATUS_TIMEOUT_MS,
-  timedOutTriggerIds,
-  toStatusMap,
   toTriggerSummaries,
   triggerBoundRoutineIds,
-  withTriggerTimeouts,
 } from "./routine-trigger-maps";
+import { useTriggerStatusTimeouts } from "./use-trigger-status-timeouts";
 
 /** Exactly the trigger props `RoutinesGrid` takes, plus the capability gate. */
 export interface TriggerSurface {
@@ -35,9 +32,9 @@ export interface TriggerSurface {
  *
  * The FETCH stays with the caller and everything downstream of it lives here:
  * a team's cross-agent list (`team-routines/use-team-trigger-statuses.ts`) fans
- * status out per agent and hands it over. Copying the timeout rule into a
- * second surface would be the one duplication we cannot afford: it is the only
- * reason a trigger row cannot lie.
+ * status out per agent and hands it over. The timeout rule itself lives in
+ * `useTriggerStatusTimeouts`, shared with the routine screen's activation chip:
+ * it is the only reason a trigger row cannot lie, so there is exactly one copy.
  *
  * The contract, and the reason this works for a merged list: `routines[i].id`
  * and `statusItems[j].routine_id` are both the GRID's row id. The team list
@@ -73,48 +70,7 @@ export function useTriggerStatusViewModel(
     setViewMode(INTEGRATIONS_VIEW_ID);
   }, [setViewMode]);
 
-  // A trigger routine that never gets a status item would otherwise say
-  // "verifying" forever (an older host, a create that silently failed). Track
-  // when each first appears WITHOUT a status, and once that has lasted past the
-  // timeout, synthesize a concrete error so the row/chip stop spinning.
-  const firstSeenRef = useRef<Record<string, number>>({});
-  const [timeoutTick, setTimeoutTick] = useState(0);
-
-  useEffect(() => {
-    const now = Date.now();
-    const ids = new Set(triggerRowIds);
-    const known = new Set((statusItems ?? []).map((i) => i.routine_id));
-    const seen = firstSeenRef.current;
-    for (const id of triggerRowIds) {
-      if (!known.has(id) && seen[id] === undefined) seen[id] = now;
-    }
-    for (const id of Object.keys(seen)) {
-      if (!ids.has(id) || known.has(id)) delete seen[id];
-    }
-    // Force one re-evaluation at the earliest pending timeout, so a routine
-    // that goes quiet still flips to the error copy even if no poll lands then.
-    const waits = Object.values(seen)
-      .map((seenAt) => seenAt + TRIGGER_STATUS_TIMEOUT_MS - now)
-      .filter((ms) => ms > 0);
-    if (waits.length === 0) return;
-    const timer = setTimeout(
-      () => setTimeoutTick((n) => n + 1),
-      Math.min(...waits),
-    );
-    return () => clearTimeout(timer);
-  }, [triggerRowIds, statusItems]);
-
-  // biome-ignore lint/correctness/useExhaustiveDependencies: timeoutTick is a deliberate recompute trigger — the body reads firstSeenRef.current + Date.now() live, and the timer bumps timeoutTick so an elapsed timeout flips the routine to the error copy.
-  const triggerStatuses = useMemo(() => {
-    const base = toStatusMap(statusItems);
-    const timedOut = timedOutTriggerIds(
-      triggerRowIds,
-      statusItems,
-      firstSeenRef.current,
-      Date.now(),
-    );
-    return withTriggerTimeouts(base, timedOut, t("trigger.statusTimeout"));
-  }, [statusItems, triggerRowIds, timeoutTick, t]);
+  const triggerStatuses = useTriggerStatusTimeouts(triggerRowIds, statusItems);
 
   const triggerSummaries = useMemo(() => {
     const bySlug = new Map(
